Change thread and reply models
Warning: I had to rewrite the migrations.
All threads and replies will be lost.

I deleted the base model for threads and replies.
Thread body and attachments were moved to the first reply.

This changes allow to simplify the logic.
Now we don't have to write the same code for threads and replies.